move compatibility checker exception for making it compatible with exomiser!
