+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/pedigree/CompatibilityCheckerException.java
@@ -1,8 +1,6 @@
-package de.charite.compbio.jannovar.pedigree.compatibilitychecker;
+package de.charite.compbio.jannovar.pedigree;
 
 import de.charite.compbio.jannovar.JannovarException;
-import de.charite.compbio.jannovar.pedigree.GenotypeList;
-import de.charite.compbio.jannovar.pedigree.Pedigree;
 
 /**
  * Exception that occurs when using invalid {@link GenotypeList} or {@link Pedigree} objects in the compatibility
