+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/pedigree/PedigreeDiseaseCompatibilityDecorator.java
@@ -1,6 +1,5 @@
 package de.charite.compbio.jannovar.pedigree;
 
-import de.charite.compbio.jannovar.pedigree.compatibilitychecker.CompatibilityCheckerException;
 import de.charite.compbio.jannovar.pedigree.compatibilitychecker.ad.CompatibilityCheckerAutosomalDominant;
 import de.charite.compbio.jannovar.pedigree.compatibilitychecker.ar.CompatibilityCheckerAutosomalRecessive;
 import de.charite.compbio.jannovar.pedigree.compatibilitychecker.xd.CompatibilityCheckerXDominant;
