+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/pedigree/compatibilitychecker/ICompatibilityChecker.java
@@ -1,5 +1,7 @@
 package de.charite.compbio.jannovar.pedigree.compatibilitychecker;
 
+import de.charite.compbio.jannovar.pedigree.CompatibilityCheckerException;
+
 public interface ICompatibilityChecker {
 
 	/**
