+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/pedigree/compatibilitychecker/ar/CompatibilityCheckerAutosomalRecessive.java
@@ -1,9 +1,9 @@
 package de.charite.compbio.jannovar.pedigree.compatibilitychecker.ar;
 
+import de.charite.compbio.jannovar.pedigree.CompatibilityCheckerException;
 import de.charite.compbio.jannovar.pedigree.GenotypeList;
 import de.charite.compbio.jannovar.pedigree.Pedigree;
 import de.charite.compbio.jannovar.pedigree.compatibilitychecker.ACompatibilityChecker;
-import de.charite.compbio.jannovar.pedigree.compatibilitychecker.CompatibilityCheckerException;
 
 /**
  * Helper class for checking a {@link GenotypeList} for compatibility with a {@link Pedigree} and autosomal recessive
