+++ jannovar-core/src/test/java/de/charite/compbio/jannovar/pedigree/CompatibilityCheckerXDominantSingletonMaleTest.java
@@ -6,8 +6,6 @@ import org.junit.Test;
 
 import com.google.common.collect.ImmutableList;
 
-import de.charite.compbio.jannovar.pedigree.compatibilitychecker.CompatibilityCheckerException;
-
 public class CompatibilityCheckerXDominantSingletonMaleTest extends CompatibilityCheckerTestBase {
 
 	@Before
