+++ jannovar-core/src/test/java/de/charite/compbio/jannovar/pedigree/CompatibilityCheckerXRecessiveSingletonMaleTest.java
@@ -6,8 +6,6 @@ import org.junit.Test;
 
 import com.google.common.collect.ImmutableList;
 
-import de.charite.compbio.jannovar.pedigree.compatibilitychecker.CompatibilityCheckerException;
-
 public class CompatibilityCheckerXRecessiveSingletonMaleTest extends CompatibilityCheckerTestBase {
 
 	@Before
